requirements.txt
# Core dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0

# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Machine Learning and Data Science
torch==2.6.0
torchvision==0.21.0
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.7.0
scipy==1.13.0

# Graph Neural Networks
torch-geometric==2.4.0
dgl==1.1.3

# Transformers and NLP
transformers==4.36.2
sentence-transformers==2.2.2
tokenizers==0.15.0

# Computer Vision and Multi-modal
opencv-python==4.8.1.78
pillow==10.1.0
clip-by-openai==1.0

# Vector Database and Similarity Search
faiss-cpu==1.7.4
pymilvus==2.3.4

# Knowledge Graph
neo4j==5.15.0
py2neo==2021.2.3

# Feature Store
feast==0.36.1

# Streaming and Event Processing
kafka-python==2.0.2
apache-flink==1.18.1

# RAG and LLM
openai==1.3.7
langchain==0.1.0
langchain-openai==0.0.2

# Reinforcement Learning
stable-baselines3==2.1.0
gymnasium==0.29.1

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==8.4.1
pytest-asyncio==0.24.0
pytest-cov==4.1.0
httpx==0.27.0

# Development and Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
tqdm==4.66.1

# Database and Caching
redis==5.0.1
alembic==1.13.1

# HTTP and API
requests==2.31.0
aiohttp==3.9.1

# Data Processing
apache-beam==2.50.0
pyspark==3.5.0

# GPU Acceleration (optional)
# nvidia-ml-py==11.525.84

# Security
cryptography==41.0.8
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Configuration Management
pyyaml==6.0.1
toml==0.10.2

# Performance and Optimization
numba==0.58.1
cython==3.0.6

# Visualization (optional)
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Jupyter for development (optional)
jupyter==1.0.0
ipykernel==6.27.1

# Backup and Recovery System
schedule==1.2.0
boto3==1.34.0
google-cloud-storage==2.10.0
azure-storage-blob==12.19.0 